Upated Simon Puzzle
Changed puzzle so it doesn't take player input while displaying pattern or showing player is wrong. Also changed it so that when a player clicks the wrong button they know immediately that it was wrong. Changed puzzle to be more forgiving. Now if player is wrong it resets to the start of the stage they got wrong instead of the start of the puzzle.